fix: resolve all clippy warnings (15 lint errors)
- Fix inconsistent digit grouping in export.rs
- Remove needless return in ai/commands.rs
- Box Ctranslate2Engine variant to reduce enum size
- Use .cloned() instead of .map(PathBuf::clone)
- Add Default impls for DiarizationEngine and TranscriptionPipeline
- Combine identical if-blocks in transcription/merge.rs
- Use &Path instead of &PathBuf in settings/config.rs
- Remove redundant bool comparison in video/processor.rs
- Add #[allow] for too-many-arguments on Tauri IPC commands

+++ src-tauri/src/commands/video_transcription.rs
@@ -66,6 +66,7 @@ pub struct VideoTranscriptionState {
 }
 
 #[tauri::command]
+#[allow(clippy::too_many_arguments)]
 pub async fn transcribe_video(
     video_path: String,
     language: Option<String>,
@@ -601,7 +602,7 @@ fn assign_block_speaker(start: f64, end: f64, turns: &[SpeakerTurn]) -> Option<S
             continue;
         }
         let dist = (t.start - start).abs();
-        let cand = (ov, -dist, -t.start as f64, idx);
+        let cand = (ov, -dist, -t.start, idx);
         let better = match &best {
             None => true,
             Some(b) => (cand.0, cand.1, cand.2, cand.3) > (b.0, b.1, b.2, b.3),
